+++ .github/workflows/check.yml
@@ -108,13 +108,13 @@ jobs:
         run: chmod +x ./gradlew
 
       - name: Run JVM tests
-        run: ./gradlew jvmTest
+        run: ./gradlew koverJvmTest
 
       - name: Upload JVM test results
         uses: actions/upload-artifact@v4
         with:
-          name: coverage-report-jvm
-          path: composeApp/build/outputs/code_coverage/
+          name: kover-report-jvm
+          path: composeApp/build/reports/
           retention-days: 1
 
   # Run wasmJs tests
@@ -324,6 +324,11 @@ jobs:
         with:
           pattern: coverage-report-*
           path: composeApp/build/outputs/code_coverage/
+
+      - name: Download JVM test results
+        uses: actions/download-artifact@v4
+        with:
+          name: kover-report-jvm
 
       - name: Set up JDK 17
         uses: actions/setup-java@v4
