+++ composeApp/pre-compile.gradle.kts
@@ -130,3 +130,93 @@ tasks
   .configureEach { dependsOn(generateSecretsTask) }
 
 tasks.matching { it.name == "clean" }.configureEach { dependsOn(cleanSecretsTask) }
+
+
+// Task to apply Supabase SQL functions
+// Gradle cache is used to avoid reapplying functions if they haven't changed
+val applySupabaseFunctions by
+tasks.registering {
+  val sqlDir = file("../supabase/functions")
+
+  // Tell Gradle: these files are inputs
+  inputs.files(sqlDir.listFiles() ?: emptyArray<File>())
+
+  // Tell Gradle: this file is the "result" marker
+  val outputMarker = layout.buildDirectory.file(".supabaseFunctionsApplied")
+  outputs.file(outputMarker)
+  val supabaseDbLink = createSupabaseDbLink()
+
+  doLast {
+    if (supabaseDbLink.isEmpty()) {
+      logger.error(
+        "Could not create Supabase database link. Please check your local.properties file."
+      )
+      return@doLast
+    }
+
+    val sqlFiles = sqlDir.listFiles()?.sortedBy { it.name } ?: emptyList()
+
+    if (sqlFiles.isEmpty()) {
+      logger.lifecycle("No SQL files found in the directory: ${sqlDir.absolutePath}")
+      return@doLast
+    }
+
+    var allSucceeded = true
+
+    sqlFiles.forEach { sqlFile ->
+      val processBuilder =
+        ProcessBuilder("psql", "-f", sqlFile.absolutePath, supabaseDbLink).apply {
+          directory(sqlDir)
+          redirectErrorStream(false)
+        }
+
+      val process = processBuilder.start()
+      val exitCode = process.waitFor()
+
+      if (exitCode != 0) {
+        logger.warn("Failed to apply: ${sqlFile.name} (exit code: $exitCode)")
+        allSucceeded = false
+      }
+    }
+
+    if (allSucceeded) {
+      outputMarker.get().asFile.writeText("Last applied at ${System.currentTimeMillis()}")
+      logger.lifecycle("Finished applying SQL functions to database.")
+    } else {
+      logger.warn("Some SQL files failed to apply.")
+    }
+  }
+}
+
+private fun createSupabaseDbLink(): String {
+  val localProperties =
+    Properties().apply {
+      val file = rootProject.file("local.properties")
+      if (file.exists()) {
+        file.inputStream().use { load(it) }
+      } else {
+        logger.error("local.properties file not found.")
+        return ""
+      }
+    }
+  val host = localProperties.getProperty(".supabase_db_host")
+  val port = localProperties.getProperty(".supabase_db_port")
+  val dbName = localProperties.getProperty(".supabase_db_name")
+  val user = localProperties.getProperty(".supabase_db_user")
+  val password = localProperties.getProperty(".supabase_db_password")
+  if (
+    host.isNullOrEmpty() ||
+    port.isNullOrEmpty() ||
+    dbName.isNullOrEmpty() ||
+    user.isNullOrEmpty() ||
+    password.isNullOrEmpty()
+  ) {
+    logger.error("One or more Supabase database properties are missing in local.properties.")
+    return ""
+  }
+  return "postgresql://$user:$password@$host:$port/$dbName"
+}
+
+tasks
+  .matching { it.name.contains("compile", ignoreCase = true) }
+  .configureEach { dependsOn(applySupabaseFunctions) }
