+++ gradle/libs.versions.toml
@@ -29,6 +29,7 @@ material3-adaptive = "1.1.2"
 composeHotReload = "1.0.0-beta04"
 logback = "1.5.16"
 slf4jApi = "2.0.16"
+kover = "0.8.1"
 
 [libraries]
 androidx-ui-test-junit4-android = { module = "androidx.compose.ui:ui-test-junit4-android", version.ref = "uiTestJunit4Android" }
@@ -76,3 +77,4 @@ room = { id = "androidx.room", version.ref = "androidx-room" }
 ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 sonar = { id = "org.sonarqube", version.ref = "sonar" }
 composeHotReload = { id = "org.jetbrains.compose.hot-reload", version.ref = "composeHotReload"}
+kover = { id = "org.jetbrains.kotlinx.kover", version.ref = "kover" }
